+++ src/Command.php
@@ -4,13 +4,10 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Sqlite;
 
-use PDOException;
 use Throwable;
 use Yiisoft\Db\Driver\Pdo\AbstractPdoCommand;
-use Yiisoft\Db\Exception\ConvertException;
 use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Exception\InvalidArgumentException;
-use Yiisoft\Db\QueryBuilder\QueryBuilderInterface;
 
 use function array_pop;
 use function count;
@@ -44,7 +41,6 @@ final class Command extends AbstractPdoCommand
                 continue;
             }
 
-            /** @psalm-var mixed */
             $result[$name] = $columns[$name] ?? $tableSchema?->getColumn($name)?->getDefaultValue();
         }
 
@@ -58,11 +54,6 @@ final class Command extends AbstractPdoCommand
         SQL;
 
         return $this->setSql($sql)->queryColumn();
-    }
-
-    protected function getQueryBuilder(): QueryBuilderInterface
-    {
-        return $this->db->getQueryBuilder();
     }
 
     /**
@@ -80,7 +71,6 @@ final class Command extends AbstractPdoCommand
     {
         $sql = $this->getSql();
 
-        /** @psalm-var array<string, string> $params */
         $params = $this->params;
 
         $statements = $this->splitStatements($sql, $params);
@@ -91,11 +81,8 @@ final class Command extends AbstractPdoCommand
 
         $result = 0;
 
-        /** @psalm-var array<array-key, array<array-key, string|array>> $statements */
         foreach ($statements as $statement) {
             [$statementSql, $statementParams] = $statement;
-            $statementSql = is_string($statementSql) ? $statementSql : '';
-            $statementParams = is_array($statementParams) ? $statementParams : [];
             $this->setSql($statementSql)->bindValues($statementParams);
             $result = parent::execute();
         }
@@ -103,43 +90,6 @@ final class Command extends AbstractPdoCommand
         $this->setSql($sql)->bindValues($params);
 
         return $result;
-    }
-
-    /**
-     * @psalm-suppress UnusedClosureParam
-     *
-     * @throws Throwable
-     */
-    protected function internalExecute(string|null $rawSql): void
-    {
-        $attempt = 0;
-
-        while (true) {
-            try {
-                if (
-                    ++$attempt === 1
-                    && $this->isolationLevel !== null
-                    && $this->db->getTransaction() === null
-                ) {
-                    $this->db->transaction(
-                        function () use ($rawSql): void {
-                            $this->internalExecute($rawSql);
-                        },
-                        $this->isolationLevel,
-                    );
-                } else {
-                    $this->pdoStatement?->execute();
-                }
-                break;
-            } catch (PDOException $e) {
-                $rawSql = $rawSql ?: $this->getRawSql();
-                $e = (new ConvertException($e, $rawSql))->run();
-
-                if ($this->retryHandler === null || !($this->retryHandler)($e, $attempt)) {
-                    throw $e;
-                }
-            }
-        }
     }
 
     /**
@@ -156,7 +106,6 @@ final class Command extends AbstractPdoCommand
     {
         $sql = $this->getSql();
 
-        /** @psalm-var array<string, string> $params */
         $params = $this->params;
 
         $statements = $this->splitStatements($sql, $params);
@@ -167,14 +116,7 @@ final class Command extends AbstractPdoCommand
 
         [$lastStatementSql, $lastStatementParams] = array_pop($statements);
 
-        /**
-         * @psalm-var array<array-key, array> $statements
-         */
         foreach ($statements as $statement) {
-            /**
-             * @psalm-var string $statementSql
-             * @psalm-var array $statementParams
-             */
             [$statementSql, $statementParams] = $statement;
             $this->setSql($statementSql)->bindValues($statementParams);
             parent::execute();
@@ -182,7 +124,6 @@ final class Command extends AbstractPdoCommand
 
         $this->setSql($lastStatementSql)->bindValues($lastStatementParams);
 
-        /** @psalm-var string $result */
         $result = parent::queryInternal($queryMode);
 
         $this->setSql($sql)->bindValues($params);
@@ -199,8 +140,6 @@ final class Command extends AbstractPdoCommand
      * @throws InvalidArgumentException
      *
      * @return array|bool List of SQL statements or `false` if there's a single statement.
-     *
-     * @psalm-param array<string, string> $params
      *
      * @psalm-return false|list<array{0: string, 1: array}>
      */
@@ -231,8 +170,6 @@ final class Command extends AbstractPdoCommand
 
     /**
      * Returns named bindings used in the specified statement token.
-     *
-     * @psalm-param array<string, string> $params
      */
     private function extractUsedParams(SqlToken $statement, array $params): array
     {